[NAYB-93] test: `UserFixture`, `user/support` 패키지로 이동
