+++ src/test/java/com/prgrms/nabmart/domain/cart/CartItemTest.java
@@ -9,10 +9,10 @@ import com.prgrms.nabmart.domain.category.SubCategory;
 import com.prgrms.nabmart.domain.category.fixture.CategoryFixture;
 import com.prgrms.nabmart.domain.item.domain.Item;
 import com.prgrms.nabmart.domain.user.User;
+import com.prgrms.nabmart.domain.user.support.UserFixture;
 import com.prgrms.nabmart.global.fixture.CartFixture;
 import com.prgrms.nabmart.global.fixture.CartItemFixture;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
-import com.prgrms.nabmart.global.fixture.UserFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
