+++ src/test/java/com/prgrms/nabmart/domain/cart/service/CartItemServiceTest.java
@@ -19,10 +19,10 @@ import com.prgrms.nabmart.domain.item.domain.Item;
 import com.prgrms.nabmart.domain.item.repository.ItemRepository;
 import com.prgrms.nabmart.domain.user.User;
 import com.prgrms.nabmart.domain.user.repository.UserRepository;
+import com.prgrms.nabmart.domain.user.support.UserFixture;
 import com.prgrms.nabmart.global.fixture.CartFixture;
 import com.prgrms.nabmart.global.fixture.CartItemFixture;
 import com.prgrms.nabmart.global.fixture.ItemFixture;
-import com.prgrms.nabmart.global.fixture.UserFixture;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
